+++ crates/extension/src/extension.rs
@@ -8,7 +8,7 @@ mod types;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
-use ::lsp::LanguageServerName;
+use ::lsp::{LanguageServerBinaryOptions, LanguageServerName};
 use anyhow::{Context as _, Result, bail};
 use async_trait::async_trait;
 use gpui::{App, Task};
@@ -68,6 +68,7 @@ pub trait Extension: Send + Sync + 'static {
         &self,
         language_server_id: LanguageServerName,
         language_name: LanguageName,
+        binary_options: LanguageServerBinaryOptions,
         worktree: Arc<dyn WorktreeDelegate>,
     ) -> Result<Command>;
 
@@ -159,6 +160,7 @@ pub trait Extension: Send + Sync + 'static {
     async fn get_dap_binary(
         &self,
         dap_name: Arc<str>,
+        allow_binary_download: bool,
         config: DebugTaskDefinition,
         user_installed_path: Option<PathBuf>,
         worktree: Arc<dyn WorktreeDelegate>,